+++ dev/apps/visibility/findMoreThan12.cpp
@@ -93,7 +93,7 @@ int main(int argc, char *argv[])
   // get the ephemeris source(s)
   EphReader ephReader;
   ephReader.verboseLevel = verbose;
-  for (int i=0; i<ephFiles.getCount(); i++)
+  for (size_t i=0; i<ephFiles.getCount(); i++)
       ephReader.read(ephFiles.getValue()[i]);
   
   gpstk::XvtStore<SatID>& ephStore = *ephReader.eph;
